+++ src/task/moph-alert.ts
@@ -12,6 +12,7 @@ import {
 const dbConnection = require('../plugins/db');
 let db: Knex = dbConnection('HIS');
 const hospcode = process.env.HOSPCODE || '';
+const limitRow = 100;
 
 // Initialize cache database on module load
 let cacheInitialized = false;
@@ -47,7 +48,6 @@ async function opdVisit(date: any = null) {
   console.log(moment().format('HH:mm:ss'), 'MOPH Alert survey: Total rows to process for date', date, ':', totalRows);
   let times = 0;
   let startRow = 0;
-  const limitRow = 100;
   let sentResult = [];
   do {
     const result = await getAndSend(date, startRow, limitRow);
@@ -59,7 +59,7 @@ async function opdVisit(date: any = null) {
   return sentResult;
 }
 
-async function getAndSend(date: any, startRow: number = -1, limitRow: number = 1000) {
+async function getAndSend(date: any, startRow: number = -1, limitRow: number = 100) {
   let rows: any = await hisModel.getVisitForMophAlert(db, date, false, startRow, limitRow);
   if (rows && rows.length > 0) {
     // Extract VNs from rows
@@ -83,8 +83,10 @@ async function getAndSend(date: any, startRow: number = -1, limitRow: number = 1
     const rowsToSend = filteredRows.map((item: any) => { return { ...item, date_service: moment(item.date_service).format('YYYY-MM-DD'), hospcode }; });
 
     // Send to MOPH
-    const result: any = await sendingToMoph('/save-moph-alert', rowsToSend);
-    console.log(moment().format('HH:mm:ss'), 'send moph alert', result.statusCode || '', result.message || '', result);
+    let result: any = await sendingToMoph('/save-moph-alert', rowsToSend);
+    console.log(moment().format('HH:mm:ss'), `send moph alert ${rowsToSend.length} rows, result status:`, result.statusCode || '', result.message || '');
+    result.resultList = result?.resultList.map(item => { delete item?.result; return item; })
+    console.log(result);
 
     // If successful, insert sent VNs into cache
     if (result.statusCode === 200) {